pci: refactor: use smallest types where possible
When working with PCI, types used to represent different parts
of the spec can use small integer types like u8/u16 instead of
usize. Usage of smaller types makes it harder to make mistakes
in the code and prevents accidental overflows.
This commit tries to replace most of the PCI used types to the
smallest ones which are enough to represent the valid range of
values required. This prevents the `try_from().unwrap()` logic
which can be error prone. Instead it only tries to widen the
type if necessary with safe `::from` functions. In minority of
cases the `#[allow(clippy::cast_possible_truncation)]` is used
with an additional `assert` or other check is used to ensure the
value is within correct range.

No functional changes intended.

+++ src/vmm/src/pci/mod.rs
@@ -35,17 +35,17 @@ pub trait PciDevice: Send {
     /// * `offset` - Offset into the register.
     fn write_config_register(
         &mut self,
-        reg_idx: usize,
-        offset: u64,
+        reg_idx: u16,
+        offset: u8,
         data: &[u8],
     ) -> Option<Arc<Barrier>>;
     /// Gets a register from the configuration space.
     /// * `reg_idx` - The index of the config register to read.
-    fn read_config_register(&mut self, reg_idx: usize) -> u32;
+    fn read_config_register(&mut self, reg_idx: u16) -> u32;
     /// Detects if a BAR is being reprogrammed.
     fn detect_bar_reprogramming(
         &mut self,
-        _reg_idx: usize,
+        _reg_idx: u16,
         _data: &[u8],
     ) -> Option<BarReprogrammingParams> {
         None
